requirements.txt
asgiref==3.8.1
bleach==6.1.0
Django==4.2.13
django-multi-captcha-admin==2.0.0
django-ranged-response==0.2.0
django-simple-captcha==0.6.0
django-summernote==0.8.20.0
django-taggit==5.0.1
mysqlclient==2.2.4
pillow==10.3.0
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
webencodings==0.5.1
